requirements.txt
altgraph==0.17.4
customtkinter==5.2.2
darkdetect==0.8.0
decorator==5.2.1
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==2.1.2
numpy==2.2.5
packaging==25.0
pillow==10.4.0
proglog==0.1.11
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
python-dotenv==1.1.0
setuptools==80.9.0
tqdm==4.67.1
yt-dlp==2025.4.30
